pyproject.toml
[project]
name = "FastAPI_and_Next_Project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn[standard]",
    "fastapi",
    "sqlalchemy",
    "psycopg2-binary",
    "pydantic",
    "passlib",
    "python-multipart",
    "alembic",
    "python-dotenv",
    "bcrypt==4.0.1",
    "passlib[bcrypt]",
    "python-jose[cryptography]"
]
